backend/requirements.txt
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==3.3.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Deprecated==1.2.14
Django==4.2.1
django-cors-headers==3.10.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
outcome==1.3.0.post0
psycopg2==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==4.0.2
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
whitenoise==5.3.0
wrapt==1.16.0
wsproto==1.2.0
zope.interface==6.0
